Removed underscores from command line arguments, and collapsed a few CLIs into common scripts. Updated readme and docs
